Compute <x> From Bare Data

Compute <x> from bare data by fitting to a constant

averx(data3pt, data2pt, ind.vec=c(1,2), ind.vec2pt=c(1,2), skip=0, t1,
t2, mps, par=c(0.6, 0.15), S=1.5, method="uwerr", nrep)

Arguments

data3pt
3pt correlator as read in with read.table from a file produced with the GWC code from routine momf2e

data2pt
2pt pseudo-scalar correlator as read in with read.table from a file produced with the GWC code from routine pssca.

ind.vec
index vector

ind.vec2pt
index vector

skip
number of measurements to be skipped from the beginning

t1
lower bound for the fitrange in time (t1,t2). Counting starts with 0.

t2
upper bound for the fitrange in time (t1,t2). Counting starts with 0.

S
passed to uwerr, see documentation of uwerr.

mps
value of the pseudo-scalar mass. If not given the value determined in averx is used

par
initial values for the fit of the pseudo-scalar mass. par[1] is the amplitude and par[2] is the mass for the cosh fit.

method
the type of error analysis to be used. Can be either “uwerr”, “boot”, “all” or “no”. For “no” (or any other string) no error analysis is performed. This might be helpful for a first impression and also to test different initial values for the fitting parameters. The latter is in particular needed for more than one state in the fit.

nrep
vector (N1, N2, ...) of replica length N1, N2. If missing it is assumed that there is only one ensemble. If there are two or more replica the parameter skip has no effect.

Value

returns a list containing

averx

value of <x>

daverx

error of <x>

data

average 3pt correlator and its error

fit.uwerr

return object of the uwerr call

mps

pseudo-scalar mass as given as input or as determined

N

number of measurements

See also

Examples

library(hadron)
# NOT RUN {
data3pt <- read.table("momf2e_op_d_0.dat")
# }# NOT RUN {
data2pt <- read.table("pssca_corr_0.dat")
# }# NOT RUN {
res3pt <- averx(data=data3pt, data2pt=data2pt, t1 = 6, t2 = 18, mps=0.13587)
# }