Computes effective masses with bootstrapping errors

Generates bootstrap samples for effective mass values computed from an object of class cf (a correlation function)

bootstrap.effectivemass(cf, boot.R, boot.l, seed=12345, type="solve", weight.factor=1)

Arguments

cf
a correlation function as an object of type cf, preferably after a call to bootstrap.cf. If the latter has not been called yet, it will be called in this function.

boot.R
only used if no preceeding call to bootstrap.cf has happened and it counts the number of bootstrap samples.

boot.l
only used if no preceeding call to bootstrap.cf has happened and it counts the bootstrap block length.

seed
the seed used in the bootstrap sample generation (for reproducability).

type
The function to be used to compute the effective mass values. Possibilities are "acosh", "solve", "log", "temporal", "shifted" and "weighted". While the first three assume normal cosh behaviour of the correlation function, "temporal" is desigend to remove an additional constant stemming from temporal states in two particle correlation functions. The same for "shifted" and "weighted", the latter for the case of two particle energies with the two particle having different energies. In the latter case only the leading polution is removed by removeTemporal.cf and taken into account here.

weight.factor
relative weight for type "weighted" only, see details

Value

An object of class effectivemass is invisibly returned. It has objects: effMass:
The computed effective mass values as a vector of length Time/2. For type="acosh" also the first value is NA, because this definition requires three time slices.

deffMass:
The computed bootstrap errors for the effective masses of the same length as effMass.

effMass.tsboot:
The boostrap samples of the effective masses as an array of dimension RxN, where R=boot.R is the number of bootstrap samples and N=(Time/2+1).

and boot.R, boot.l, Time

Details

A number of types is implemented to compute effective mass values from the correlation function:

"solve": the ratio
\(C(t+1) / C(t) = \cosh(-m*(t+1)) / \cosh(-m*t)\)
is numerically solved for m.

"acosh": the effective mass is computed from
\(m=acosh((C(t-1)+C(t+1)) / (2C(t)))\)
Note that this definition is less tolerant against noise.

"log": the effective mass is defined via
\(m=\log(C(t) / C(t+1))\)
which has artifacts of the periodicity at large t-values.

"temporal": the ratio
\([C(t)-C(t+1)] / [C(t-1)-C(t)] = [\cosh(-m*(t))-\cosh(-m*(t+1))] / [\cosh(-m*(t-1))-\cosh(-m(t))]\)
is numerically solved for \(m(t)\).

"shifted": like "temporal", but the differences \(C(t)-C(t+1)\) are assumed to be taken already at the correlator matrix level using removeTemporal.cf and hence the ratio
\([C(t+1)] / [C(t)] = [\cosh(-m*(t))-\cosh(-m*(t+1))] / [\cosh(-m*(t-1))-\cosh(-m(t))]\)
is numerically solved for \(m(t)\).

"weighted": like "shifted", but now there is an additional weight factor \(w\) from removeTemporal.cf to be taken into account, such that the ratio
\([C(t+1)] / [C(t)] = [\cosh(-m*(t))-w*\cosh(-m*(t+1))] / [\cosh(-m*(t-1))-w*\cosh(-m(t))]\)
is numerically solved for \(m(t)\) with \(w\) as input.

References

arXiv:1203.6041

See also

Examples

data(samplecf)
samplecf <- bootstrap.cf(cf=samplecf, boot.R=1500, boot.l=2, seed=1442556)

#> Error: inherits(cf, "cf_orig") ist nicht TRUE

effmass <- bootstrap.effectivemass(cf=samplecf, boot.R=1500, boot.l=2, seed=1442556)

#> Error in bootstrap.effectivemass(cf = samplecf, boot.R = 1500, boot.l = 2,     seed = 1442556): unbenutzte Argumente (boot.R = 1500, boot.l = 2, seed = 1442556)

summary(effmass)

#> Error in object[[i]]: Objekt des Typs 'closure' ist nicht indizierbar

plot(effmass, ylim=c(0.14,0.15))

#> Error: Argument "t" fehlt (ohne Standardwert)