perform a bootstrap analysis of a GEVP

perform a bootstrap analysis of a GEVP for a real, symmetric correlator matrix

bootstrap.gevp(cf, t0=1, boot.R=400, boot.l=2, matrix.size=2,
element.order=c(1,2,3,4), seed=1234, sort.type="vectors", sort.t0=TRUE)

Arguments

cf
correlation matrix obtained with a call to extrac.obs.

t0
initial time value of the GEVP, must be in between 0 and Time/2-2. Default is 1.

matrix.size
the size of the square matrix

element.order
specifies how to fit the n linearly ordered single correlators into the correlator matrix. element.order=c(1,2,3,4) leads to a matrix matrix(cf[element.order], nrow=2). Double indexing is allowed.

boot.R
number of bootstrap samples. Used only if cf was not yet bootstrapped using bootstrap.cf.

boot.l
block size for autocorrelation analysis

seed
seed for the random number generation used for boostrapping.

sort.type
Sort the eigenvalues either in descending order, or by using the scalar product of the eigenvectors with the eigenvectors at \(t=t_0+1\). Possible values are "values" or "vectors".

sort.t0
for sort.type "vectors" use \(t_0\) as reference or \(t-1\).

Value

Returns an object of class gevp with member objects:

cf:
The input data, if needed bootstrapped with bootstrap.cf.

res.gevp:
The object returned from the call to gevp. For the format see gevp.

gevp.tsboot:
The bootstrap samples of the GEVP. For the format see gevp.

References

Michael, Christopher and Teasdale, I., Nucl.Phys.B215 (1983) 433, DOI: 10.1016/0550-3213(83)90674-0
Blossier, B. et al., JHEP 0904 (2009) 094, DOI: 10.1088/1126-6708/2009/04/094, arXiv:0902.1265

See also

Examples

# NOT RUN {
pion.cor.gevp <- bootstrap.gevp(pion.cor, t0=1)
# }# NOT RUN {
## extract the first principal correlator
# }
# NOT RUN {
pion.pc1 <- gevp2cf(pion.cor.gevp, id=1)
# }# NOT RUN {
## which can now be treated like a bootstrapped correlation function
# }
# NOT RUN {
pion.pc1.effectivemass <- bootstrap.effectivemass(cf=pion.pc1, type="acosh")
# }# NOT RUN {
pion.pc1.effectivemass <- fit.effectivemass(pion.pc1.effectivemass, t1=12, t2=23, useCov=FALSE)
# }# NOT RUN {
summary(pion.pc1.effectivemass)
# }# NOT RUN {
plot(pion.pc1.effectivemass, xlab=c("t/a"), ylab=c("aM"))
# }