bootstrap an NLS fit

performs and bootstraps an non-linear least-squares fit to data with y and x errors.

bootstrap.nlsfit(fn, par.guess, errormodel = "yerrors", sim =
"parametric", boot.R, y, dy, x, dx = NULL, bsamples, useCov, ...)

summary(object, digits=2, ...)

print(x, digits=2, ...)

plot(x, ..., xlim, ylim, rep=FALSE, col.line="black",
col.band="gray", lty=c(1), lwd=c(1))

Arguments

fn
the (non-linear) function to be fitted to the data. Its first argument must be the fit parameters named par. The second must be x, the explaining variable.

par.guess
Initial guess values for the fit parameters.

errormodel
can be either "yerrors" for errors only on the y-axis, or "xyerrors" for errors on both axis.

sim
can be "parametric" for a parametric bootstrap approach. Using the values dy (and dx, if applicable), bootstrap samples are generated assuming a Gaussian error distribution.

boot.R
the number of bootstrap samples to use.

y
the data as a one-dimensional numerical vector to be described by the fit function.

dy
errors of y as a one-dimensional numerical vector.

x
Values of the explaining variable in form of a one-dimensional numerical vector.

dx
errors of x as a one-dimensional numerical vector.

bsamples
bootstrap samples of y (and x, if applicable). Must be provided as array of dimensions (npar+1, boot.R+1) with npar equals to length(par.guess) in case of 'yerrors' and For 'xyerrors' to length(par.guess) + length(x)..

useCov
perform a fully correlated fit by deterining the variance-covariance matrix from the provided bootstrap samples. This works only if bootstrap samples are provided. Otherwise, an uncorrelated fit will be performed.

…
additional parameters to be passed on to the fit function fn.

object
object returned by bootstrap.nlsfit

digits
number of significant digits to print in summary or print.

xlim
x limits of the plot.

ylim
y limits of the plot.

rep
If set to TRUE, operate like "replot" in gnuplot. Allows adding points with error bars to the current plot. Switches the underlying plotting routine from plot to points.

col.line
line colour.

col.band
error band colour.

lwd
line width for fitted curve.

lty
line type of fitted curve.

Details

Value

returns a list of class 'bootstrapfit'. It returns all input parameters and adds in addition the following:

t0

the one dimensional numerical vector of length npar+1. npar is the number of fit parameters. In case of 'yerrors' this equals length(par.guess). For 'xyerrors' this equals length(par.guess) + length(x). t0 contains the best fit parameters obtained on the original data. The last element in t0 is the chisquare value.

t

an array of dimensions (npar+1, boot.R+1) with npar as in t0. The first row is identical to t0. The other rows contain the individual bootstrap observations.

bsamples

the bootstrap samples used as an array of dimensions (npar+1, boot.R+1) with npar as in t0.

Qval

the p-value of the fit on the original data

chisqr

the residual chisqr value.

dof

the residual degrees of freedom of the fit.

nx

the number of x-values.

See also

Examples

value <- c(0.1, 0.2, 0.3)
dvalue <- c(0.01, 0.01, 0.015)
x <- c(1,2,3)
dx <- c(0.1, 0.1, 0.1)
boot.R <- 1500
## with xy-errors and no correlated fit
fitres <- bootstrap.nlsfit(fn=function(par, x) par[1] + par[2]*x, par.guess=c(1,1), errormodel="xyerrors", boot.R=boot.R, y=value, dy=dvalue, x=x, dx=dx, useCov=FALSE)

#> Lade nötiges Paket: minpack.lm

summary(fitres)

#> bootstrap nls fit
#> 
#> model xyerrors 
#>     best fit parameters with errors, bootstrap bias and 68% confidence interval
#> 
#>         par     bias   ci16  ci84
#> 1 0.000(24)   0.0010 -0.027 0.024
#> 2 0.100(12) -0.00027  0.088  0.11
#> 
#>    correlation matrix of the fit parameters
#> 
#>           X1          X2         X3          X4         X5
#> 1  1.0000000 -0.92945749 -0.6653684 -0.10991255  0.3041262
#> 2 -0.9294575  1.00000000  0.5412653 -0.09711556 -0.4477866
#> 3 -0.6653684  0.54126530  1.0000000  0.18887980 -0.1024149
#> 4 -0.1099125 -0.09711556  0.1888798  1.00000000  0.1997793
#> 5  0.3041262 -0.44778656 -0.1024149  0.19977931  1.0000000
#> 
#>  estimates for x-values with errors, bootstrap bias and 68% confidence interval
#> 
#>           x     bias ci16 ci84
#> 1  1.00(10)  -0.0011 0.90  1.1
#> 2 2.000(85) -0.00077  1.9  2.1
#> 3  3.00(10)  -0.0056  2.9  3.1
#> 
#>    chi^2 and fit quality
#> chisqr / dof = 2.234484e-15 / 1 = 2.234484e-15 
#> p-value 1 

## with covariance matrix and y-errors
bootstrapsamples <- array(NA, dim=c(boot.R+1, length(value)))
bootstrapsamples[1,] <- value
## some indices
crr <- c(1:(boot.R+1))
rr <- c(2:(boot.R+1))
bootstrapsamples[rr, 1] <- rnorm(n=boot.R, mean = value[1], sd = dvalue[1])
bootstrapsamples[rr, 2] <- rnorm(n=boot.R, mean = value[2], sd = dvalue[2])
bootstrapsamples[rr, 3] <- rnorm(n=boot.R, mean = value[3], sd = dvalue[3])

## the fit without correlation and y-errors only
fitres <- bootstrap.nlsfit(fn=function(par, x) par[1] + par[2]*x, par.guess=c(1,1), errormodel="yerrors", boot.R=boot.R, y=value, dy=dvalue, x=x, useCov=TRUE, bsamples=bootstrapsamples[rr,])
summary(fitres)

#> bootstrap nls fit
#> 
#> model yerrors 
#>     best fit parameters with errors, bootstrap bias and 68% confidence interval
#> 
#>          par     bias   ci16  ci84
#> 1  0.000(17) -0.00025 -0.017 0.017
#> 2 0.1000(89)  0.00016  0.091  0.11
#> 
#>    correlation matrix of the fit parameters
#> 
#>           X1         X2
#> 1  1.0000000 -0.9272133
#> 2 -0.9272133  1.0000000
#> 
#>    chi^2 and fit quality
#> chisqr / dof = 1.710734e-15 / 1 = 1.710734e-15 
#> p-value 1