finite size corrections a la Colangelo, Duerr, Haefeli, but re-expanded as series in the quark mass

finite size corrections a la Colangelo, Duerr, Haefeli, but re-expanded as series in the quark mass

cdhnew(rev=-1, aLamb1=0.055, aLamb2=0.58, aLamb3, aLamb4,
       ampiV, afpiV, aF0, a2B0mu, L, printit=FALSE, use.cimpl=TRUE)

Arguments

rev
\(rev=-1\) corrects from \(L\) to \(L=\infty\), \(rev=+1\) the other way around

aLamb1
The four low energy \(\Lambda_{1-4}\)constants in lattice units.

aLamb2
see aLamb1.

aLamb3
see aLamb1.

aLamb4
see aLamb1.

ampiV
pseudo scalar mass values to be corrected

afpiV
pseudo scalar decay constant values to be corrected

aF0
\(af_0\) in lattice units

a2B0mu
\(2B_0\mu\) in lattice units, where \(\mu\) is the quark mass and \(B_0\) a low energy constant

L
the lattice spatial extent

printit
if set to TRUE the corrections are printed

use.cimpl
use the four times faster direct c Implementation of the correction routine

Value

a list with the corrected values for mpi and fpi

Details

see reference for details. We use the simplyfied formulae for the S quantities, see eq. (59) in first reference.

References

Gilberto Colangelo, Stephan Durr, Christoph Haefeli, Nucl.Phys.B721:136-174,2005. hep-lat/0503014

and

R. Frezzotti, V. Lubicz, S. Simula, arXiv:0812.4042 [hep-lat]

Examples

mu <- c(0.004, 0.006, 0.008, 0.010, 0.004)
L <- c(24, 24, 24, 24, 32)
mps <- c(0.14448, 0.17261, 0.19858, 0.22276, 0.14320)
fps <- c(0.06577, 0.07169, 0.07623, 0.07924, 0.06730)
aLamb1 <- 0.05
aLamb2 <- 0.5
aLamb3 <- 0.38
aLamb4 <- 0.66
aF0    <- 0.051
a2B    <- 5.64
cdhres <- cdhnew(rev=+1, aLamb1=aLamb1, aLamb2=aLamb2, aLamb3=aLamb3,
                 aLamb4=aLamb4, ampiV=mps, afpiV=fps, aF0=aF0,
                 a2B0mu=a2B*mu, L=L, printit=TRUE)
cdhres$mpiFV

#> [1] 0.1463914 0.1738754 0.1994823 0.2234273 0.1435068

cdhres$fpiFV

#> [1] 0.06406079 0.07079489 0.07569973 0.07890527 0.06700416