Correlation function container

This function cf() creates containers for correlation functions of class cf. This class is particularly designed to deal with correlation functions emerging in statistical and quantum field theory simulations. Arithmetic operations are defined for this class in several ways, as well as concatenation and is.cf and as.cf.

cf(nrObs = 1, Time = NA, nrStypes = 0, symmetrised = FALSE)

Arguments

nrObs
Integer, number of different measurements contained in this correlation function. One can use c.cf to add multiple observables into one container. This is for instance needed when passing to the gevp function.

Time
Integer, full time extent.

nrStypes
Integer, number of smearing types.

symmetrised
Logical, indicating whether the correlation function has been symmetrized.

Details

And last but not least, these are the fields that are used somewhere in the library but we have not figured out which mixin these should belong to:

smeared: Logical, whether the correlation function has smeared data.

scf: Like cf, but with the smeared data.

iscf: Like icf, but with the smeared data.

nrSamples: TODO

conf.index: TODO

obs: TODO

N: Integer, number of measurements.

blockind: TODO

jack.boot.se: TODO

See also