Computes the pseudoscalar decay constant for the twisted mass case from the pseudoscalar amplitude and mass

From a mass and amplitude determination (using matrixfit or fit.effectivemass, bootstrap.gevp and gevp2amplitude the pseudoscalar decay constant is determined for the case of Wilson twisted mass fermions from the pseudoscalar amplitude and mass

computefps(mfit, PP, mass, mu1, mu2, Kappa, normalisation="cmi",
           disprel="continuum", boot.fit=TRUE)

Arguments

mfit
An object of type matrixfit or gevp.amplitude generated with matrixfit or gevp2amplitude, respectively.

PP
If mfit is missing this must contain the value for the pseudoscalar amplitude.

mass
If mfit is missing this must contain the value for the pseudoscalar mass.

mu1, mu2
The values for the twisted quark masses involved in the pseudoscalar meson. If mu2 is missing it will be assumed to be equal to mu1.

Kappa
The \(\kappa\)-value of the run, needed only if normalisation="cmi".

normalisation
normalisation of the correlators. If set to "cmi" the \(\kappa\) value must be specified.

disprel
One of "continuum" or "lattice". Indicates whether the formula for the decay constant should take into account the lattice dispersion relation for the meson. Theoretically this can reduce lattice artefacts for heavy mesons.

boot.fit
If set to FALSE, the computation is not bootstrapped, even if the matrixfit or gevp.amplitude contain bootstrap samples. This is a useful time-saver if error information is not strictly necessary. Of course, this affects the return values related to the bootstrap, which are set to NA.

Value

If mfit ist missing the value of fps will printed to stdout and returned as a simple numerical value.

If mfit is available, this object will be returned but with additional objects added: fps, fps.tsboot, mu1,mu2, normalistaion and Kappa if applicable.

Details

The pseudoscalar decay constant is computed from
$$f_\mathrm{PS} = 2\kappa(\mu_1+\mu_2)\frac{PP}{\sqrt{2}\sqrt{m_\mathrm{PS}}^3}$$ for normalisation="cmi" or $$f_\mathrm{PS} = (\mu_1+\mu_2)\frac{PP}{\sqrt{2}\sqrt{m_\mathrm{PS}}^3}$$ expecting physical normalisation of the amplitudes.
When disprel="lattice",
$$\sqrt{m_{\mathrm{PS}}^3}$$ is replaced with $$\sqrt{m_{\mathrm{PS}}} \sinh{m_{\mathrm{PS}}}$$ which can reduce lattice artefacts for heavy meson masses.

See also