Computes the pseudoscalar decay constant for the Osterwalder Seiler case from the pseudoscalar amplitude and mass

From a mass and amplitude determination (using matrixfit) the pseudoscalar decay constant is determined for the case of Osterwalder Seiler (OS) fermions from the AS and SS amplitude (in the twisted basis), ZA and the OS pion mass.

computefpsOS(mfit, Kappa, normalisation="cmi", boot.fit=TRUE, ZA=1, ZAboot, dZA)

Arguments

mfit
An object of type matrixfit generated with matrixfit. The correlation matrix (SS, SA, AS, AA) must have been analysed, where the correlators are in the twisted basis.

ZA
The value of the renormalisation constant \(Z_A\).

dZA
The value of the (normally distributed) error of the renormalisation constant \(Z_A\).

ZAboot
Bootstrap samples for \(Z_A\). If they are provided, they are used for computing fps, if not, bootstrap samples are generated from dZA. If both are missing, the error of \(Z_A\) is not taken into account.

Kappa
The \(\kappa\)-value of the run, needed only if normalisation="cmi".

normalisation
normalisation of the correlators. If set to "cmi" the \(\kappa\) value must be specified.

boot.fit
If set to FALSE, the computation is not bootstrapped, even if the matrixfit or gevp.amplitude contain bootstrap samples. This is a useful time-saver if error information is not strictly necessary. Of course, this affects the return values related to the bootstrap, which are set to NA.

Value

If mfit is available, this object will be returned but with additional objects added: fpsOS, fpsOS.tsboot, normalistaion, ZA, ZAboot and kappa if applicable.

Details

The pseudoscalar decay constant is computed from
$$f_\mathrm{PS}^\mathrm{OS} = Z_A \sqrt{2}\kappa\frac{\langle 0| A|\pi\rangle}{m_\mathrm{PS}}$$ for normalisation="cmi" or $$f_\mathrm{PS}^\mathrm{OS} = Z_A \frac{\langle 0| A|\pi\rangle}{m_\mathrm{PS}}$$ expecting physical normalisation of the amplitudes.

See also