Extract One or More Gamma Combinations from am CMI Correlator

Extracts one or more gamma matrix combinations (observables) from a correlator stored in cmi format

extract.obs(cmicor, vec.obs=c(1), ind.vec=c(1,2,3,4,5),
                        sym.vec, sign.vec, verbose=FALSE, symmetrise=TRUE)

Arguments

cmicor
an correlator object in cmi format

vec.obs
vector containing the numbers of observables to be extracted.

ind.vec
Index vector indexing the column numbers in cmicor to be used. The first must be the observable index, the second the smearing type index, the third the time, the fourth C(+t) and the fifth C(-t).

Index vector indexing the column numbers in cmiloop to be used. The first must be the observable index, the second the smearing type index, the third the time, the fourth ReTL, the fifth ImTL, the sixth ReTF and the seventh ImTF.

verbose
Increases verbosity of the function.

sym.vec
a vector of bools of length equal to the number of observables indicating whether C(t) is symmetric in t, i.e. whether C(+t) and C(-t) should be added or subtracted. If not given C(+t) and C(-t) will be assumed to be symmetric.

sign.vec
a sign vector of length equal to the number of observables indicating whether the corresponding correlation function should be multiplied by +-1.

symmetrise
if set to TRUE, the correlation function will be averaged for t and T-t, with the sign depending on the value of sym. Note that currently the correlator with t-values larger than T/2 will be discarded.

Value

returns a list containing

cf

for extract.obs: array containing the correlation function with dimension number of files times (nrObs*nrStypes*(T/2+1)). C(t) and C(-t) are averaged according to sym.vec. for extract.loop: ReTL

icf

for extract.loop only: ImTL

scf

for extract.loop only: ReTF

sicf

for extract.loop only: ImTF

Time

The time extend of the correlation functions.

nrStypes

The number of smearing combinations.

nrObs

The number of observables.

nrSamples

for extrac.loop only: the number of samples found in the files.

Details

C(t) and C(-t) are averaged as indicated by sym.vec.

References

See also

Examples

# NOT RUN {
cmicor <- readcmidatafiles("outprc", skip=1)
# }# NOT RUN {
cf <- extract.obs(cmicor, vec.obs=c(1,3))
# }