Fits a constant to effective mass data

Performs a correlated fit of a constant to data generated with bootstrap.effectivemass.

fit.effectivemass(cf, t1, t2, useCov = FALSE, replace.na = TRUE,
                  boot.fit = TRUE, autoproceed = FALSE)

Arguments

cf
An object of class effectivemass generated by a call to bootstrap.effectivemass.

t1, t2
The fit range. If several correlators are fitted, this is automatically replicated accordingly. The fit range is adjusted such that NAs are removed from the fit. They must fulfill \(t_1<t_2\). For symmetric correlators, they must both run from 0 to T/2-1, otherwise from 0 to T-1.

useCov
Use the correlated chisquare. This works only for not too noisy data.

replace.na
The functions inverted to determine the effective mass values might, due to fluctuations, return NA. If replace.na=TRUE, these are reaplaced in the bootstrap samples by randomly chosen values from the distribution that are not NA. Otherwise the fits in which the NA values occur will fail.

boot.fit
If set to FALSE, the effective mass fit is not bootstrapped, even though bootstrap samples are still used to estimate the variance-covariance matrix for the correlated fit. This is a useful time-saver if error information is not strictly necessary. Of course, this affects the return values related to the bootstrap, which are set to NA.

autoproceed
When the inversion of the variance-covariance matrix fails, the default behaviour is to abort the fit. Setting this to TRUE means that the fit is instead continued with a diagonal inverse of the variance-covariance matrix.

Value

An object with class effectivemassfit is returned. It contains all the data of the input object effMass with the following additional member objects:

opt.res: the object returned by the optim on the original data.

massfit.tsboot: the bootstrap values of the mass and the chisquare function.

ii: the index array of data used in the fit.

invCovMatrix: the inverse covariance matrix.

dof: the degrees of freedom of the fit.

t1,t2: the fit range.

Details

A correlated chisquare minimisation is performed on the original data as well as on all bootstrap samples generated by bootstrap.effectivemass. The inverse covariance matrix is generated as described in hep-lat/9412087 in case of too little data to relibably estimate it.

References

C.Michael, A.McKerrell, Phys.Rev. D51 (1995) 3745-3750, hep-lat/9412087

See also

Examples

data(samplecf)
samplecf <- bootstrap.cf(cf=samplecf, boot.R=1500, boot.l=2, seed=1442556)

#> Error: inherits(cf, "cf_orig") ist nicht TRUE

effmass <- fit.effectivemass(bootstrap.effectivemass(cf=samplecf, boot.R=1500, boot.l=2, seed=1442556), t1=15, t2=23)

#> Error in bootstrap.effectivemass(cf = samplecf, boot.R = 1500, boot.l = 2,     seed = 1442556): unbenutzte Argumente (boot.R = 1500, boot.l = 2, seed = 1442556)

summary(effmass)

#> Error in object[[i]]: Objekt des Typs 'closure' ist nicht indizierbar

plot(effmass, ylim=c(0.14,0.15))

#> Error in invcosh((data[1] + data[4])/(data[2] + data[3]), timeextent = timeextent,     t = t): Argument "t" fehlt (ohne Standardwert)