Extracts physical amplitudes from a GEVP

Given a GEVP generated with bootstrap.gevp and masses determined from the principle correlator with given id, the physical amplitudes are extracted and bootstraped. The man amplitude is determined from a constant fit to the data in the specified time range.

gevp2amplitude(gevp, mass, id=1, op.id=1, type="cosh", t1, t2,
               useCov=TRUE, fit=TRUE)

Arguments

gevp
An object of class gevp as generated with a call to bootstrap.gevp.

mass
Optimally, this is an object either of class effectivemassfit generated using fit.effectivemass or of class matrixfit generated with matrixfit to the principal correlator extracted using gevp2cf applied to gevp using the same value of id.

It can also be given as a numerical vector with the bootstrap samples as entries. The mean will then be computed as the bootstrap mean over this vector. The number of samples must agree with the number of bootstrap samples in gevp.

id
The index of the principal correlator to extract, i.e. the physical state to extract.

op.id
The index of the operator for which to extract the amplitude.

type
The symmetry of the pricipal correlator in time, can be either "cosh" or "sinh".

t1,t2
The time range in which to fit the amplitude starting with 0. If not given it will be tried to infer these from the mass object.

useCov
Use the covariance matrix for fitting the constant to the amplitude data.

fit
perform a fit to the data.

Value

Details

See also

Examples

# NOT RUN {
## apply a GEVP analysis
# }
# NOT RUN {
pion.cor.gevp <- bootstrap.gevp(pion.cor, t0=1)
# }# NOT RUN {
## extract the first principal correlator
# }
# NOT RUN {
pion.pc1 <- gevp2cf(pion.cor.gevp, id=1)
# }# NOT RUN {
pion.pc1.effectivemass <- bootstrap.effectivemass(cf=pion.pc1, type="acosh")
# }# NOT RUN {
pion.pc1.effectivemass <- fit.effectivemass(pion.pc1.effectivemass, t1=12, t2=23, useCov=TRUE)
# }# NOT RUN {
## now determine the amplitude
# }
# NOT RUN {
pion.pc1.amplitude <- gevp2amplitude(pion.cor.gevp, pion.pc1.effectivemass)
# }# NOT RUN {
## compute also the pion decay constant
# }
# NOT RUN {
pion.pc1.amplitude <- computefps(pion.pc1.amplitude, Kappa=0.125, mu1=0.003)
# }# NOT RUN {
summary(pion.pc1.amplitude)
# }# NOT RUN {
plot(pion.pc1.amplitude)
# }