determines pion mass and pcac mass from online measured correlator of the HMC code

determines pion mass and pcac mass from online measured correlator of the HMC code

onlinemeas(data, t1, t2, stat_range, S=1.5, pl=FALSE, skip=0,
           iobs=1, ind.vec=c(1,3,4,5), mu=0.1, kappa=0.125,
           boot.R=99, boot.l=10, tsboot.sim="geom",
           method="uwerr", fit.routine="optim", nrep,
           oldnorm = FALSE)

Arguments

data
data to be fitted to as e.g. the output of readcmicor. Currently only cmicor format is supported.

t1
lower bound for the fitrange in time (t1,t2). Counting starts with 0.

t2
upper bound for the fitrange in time (t1,t2). Counting starts with 0.

S
passed to uwerr, see documentation of uwerr.

pl
logical: if set to TRUE the function produces plots

stat_range
range of data to be included in the analysis.

skip
number of measurements to be discarded at the beginning of the time series. skip has no effect if two or more replica are used, see argument nrep.

iobs
if there are several operators available (local-local, local-smeared, etc.), then this labels these (for cmi format)

ind.vec
index vector indexing the column numbers in cmicor to be used

boot.R
number of bootstrap samples for bootstrap analysis

boot.l
average block size for blocking analysis with tsboot

tsboot.sim
The type of simulation required to generate the replicate time series. See tsboot for details.

method
the type of error analysis to be used. Can be either “uwerr”, “boot”, “all” or “no”. For “no” (or any other string) no error analysis is performed. This might be helpful for a first impression and also to test different initial values for the fitting parameters. The latter is in particular needed for more than one state in the fit.

nrep
vector (N1, N2, ...) of replica length N1, N2. If missing it is assumed that there is only one ensemble. If there are two or more replica the parameter skip has no effect.

fit.routine
The fit routine to be used. Default is “gsl”, which uses the gnu scientific library “gsl_multifit_fdfsolver” solver to minimise the chisquare. All other values lead to the usage of R's optim function. The latter choice might be significantly slower.

oldnorm
If set to “TRUE”, the old online measurement normalisation of “tmLQCD” prior to version 5.2.0 is used in order to get correct values for the pion decay constant.

Value

returns an object of class ofit with the following items

fitresult

result from the fit as returned by optim

fitresultpp

Fit result of the PP correlator only

t1

lower bound for the fitrange in time (t1,t2). Counting starts with 0.

t2

upper bound for the fitrange in time (t1,t2). Counting starts with 0.

N

number of measurements found in the data

Time

Time extent found in the data

fitdata

data.frame containing the time values used in the fit, the averaged correlator and its error and the value of Chi for each time value

uwerrresultmps

the result of the time series analysis for the lowest mass as carried out by uwerr

uwerrresultmpcac

the result of the time series analysis for the PCAC mass carried out by uwerr, see details

effmass

effective masses in the pion channel

matrix.size

size of the data matrix, copied from input

boot

object returned by the call to boot if method was set correspodingly. Otherwise NULL.

tsboot

object returned by the call to tsboot if method was set correspodingly. Otherwise NULL.

method

error analysis method as copied from input

fit.routine

fit.routine as copied from input

nrep

nrep as copied from input

dpaopp

data.frame containing the pcac masses computed not with a fit, but with the derivative method for all time values in between t1 and t2

Details

The online measurements in the HMC code compute the PP and PA correlation functions summed over spatial x for all t. We analyse these correlators in different ways:

First, only the PP correlator is analysed and fitted by \(p_1^2\cosh(-m(t-T/2))\) for \(m\) and \(p_1\).

Second, PP and PA correlators are fitted together with three parameters as \(C_\mathrm{PP} = p_1^2\cosh(-m(t-T/2))\) and \(C_\mathrm{PA} = p_1p_2\cosh(-m(t-T/2))\) in a simultaneous fit. \(m\) is then the pseudo scalar mass and the pcac mass is determined from $$m_\mathrm{PCAC} = m_\mathrm{PS} \frac{p_2}{2p_1}$$

Finally, the PCAC mass can also be determined computing $$m_\mathrm{PCAC}(t) =% \frac{C_\mathrm{PA}(t+1)-C_\mathrm{PA}(t-1)}{4C_\mathrm{PP}(t)}$$ using the symmetric finite difference operator.

See also