Fit Pseudo Scalar Sector

fits one or several cosh and sinh to data for extracting the pseudo-scalar meson mass, excited states in the pseudo-scalar channel and determines decays constant and \(m_\mathrm{PCAC}\)

This is so far only for twisted mass lattice QCD in the twisted basis.

pion(cmicor, mu=0.1, kappa=0.156, t1, t2, S=1.5, pl=FALSE, skip=0,
       variational=list(ta=3, tb=4, N=6), ind.vec=c(1,3,4,5),
       no.masses=1, matrix.size=2, boot.R=99, boot.l=10, tsboot.sim="geom",
       method="uwerr", fit.routine="optim", mass.guess, par.guess, nrep)

Arguments

cmicor
data to be fitted to as e.g. the output of readcmicor

mu
the value of the bare quark twisted mass

kappa
the value of the hopping parameter

t1
lower bound for the fitrange in time (t1,t2). Counting starts with 0.

t2
upper bound for the fitrange in time (t1,t2). Counting starts with 0.

S
passed to uwerr, see documentation of uwerr.

pl
logical: if set to TRUE the function produces plots

skip
number of measurements to be discarded at the beginning of the time series. skip has no effect if two or more replica are used, see argument nrep.

variational
list of parameters used for the variational analysis

ind.vec
index vector indexing the column numbers in cmicor to be used

no.masses
number of masses to be extracted. This argument will set the number of exponentials to be used in the fit.

Note that this is not yet stable for no.masses > 1.

matrix.size
matrix size to be used in the fit. Can be currently set to 2,4 and 6.

boot.R
number of bootstrap samples for bootstrap analysis

boot.l
average block size for blocking analysis with tsboot

tsboot.sim
The type of simulation required to generate the replicate time series. See tsboot for details.

method
the type of error analysis to be used. Can be either “uwerr”, “boot”, “all” or “no”. For “no” (or any other string) no error analysis is performed. This might be helpful for a first impression and also to test different initial values for the fitting parameters. The latter is in particular needed for more than one state in the fit.

mass.guess
numerical vector of mass-values to be used as initial values in the fit. If given, it must have at least no.masses entries.

par.guess
numerical vector of couling parameter values to be used as initial values in the fit. The order is \(P_L\), \(P_F\), \(A_L\), \(A_F\), \(4_L\), \(4_F\), for notation see below. If given, it must have at least no.masses times matrix.size entries.

nrep
vector (N1, N2, ...) of replica length N1, N2. If missing it is assumed that there is only one ensemble. If there are two or more replica the parameter skip has no effect.

fit.routine
The fit routine to be used. Default is “gsl”, which uses the gnu scientific library “gsl_multifit_fdfsolver” solver to minimise the chisquare. All other values lead to the usage of R's optim function. The latter choice might be significantly slower.

Value

returns an object of class pionfit with the following items

fitresult

result from the fit as returned by optim

t1

lower bound for the fitrange in time (t1,t2). Counting starts with 0.

t2

upper bound for the fitrange in time (t1,t2). Counting starts with 0.

N

number of measurements found in the data

Time

Time extent found in the data

fitdata

data.frame containing the time values used in the fit, the averaged correlator and its error and the value of Chi for each time value

uwerrresultmps

the result of the time series analysis for the lowest mass as carried out by uwerr

uwerrresultmps2

the result of the time series analysis for the second lowest mass as carried out by uwerr if no.masses larger than 1.

uwerrresultmps3

the result of the time series analysis for the second lowest mass as carried out by uwerr, if no.masses larger than 2.

uwerrresultfps

the result of the time series analysis for the pseudo-scalar decay constant as carried out by uwerr.

uwerrresultmpcac

the result of the time series analysis for the PCAC mass as carried out by uwerr if no.masses larger than 1.

uwerrresultzv

the gamma method analysis for \(Z_V\), if matrix size equals 6

effmass

effective masses

mu

the value of the bare quark twisted mass

kappa

the value of the hopping parameter

variational.masses

mass values as determined by the variational analysis

no.masses

no.masses determined, copied from input

matrix.size

size of the data matrix, copied from input

boot

object returned by the call to boot if method was set correspodingly. Otherwise NULL.

tsboot

object returned by the call to tsboot if method was set correspodingly. Otherwise NULL.

var.res

the full result of the variational analysis, as returned by a call to variational

method

error analysis method as copied from input

fit.routine

fit.routine as copied from input

nrep

nrep as copied from input

Details

A call to pion returns an object, for which several generic functions are available, such as summary and plot. The function should always be used by also considering the output of plot. In particular, in case of the gamma matrix method it should always be checked whether the autocorrelation function could be integrated in a sensible way. For this reason the autocorrelation function is plotted as well as the integrated autocorrelation time as a function of the cut-off value for the integral. Obviously, the autocorrelation function should be zero at the cut-off and for larger values only fluctuate around zero. In other words, the autocorrelation time should reach a plateau. Changing the parameter S (see above) allows to influence the automatic determination of the cut-off value (see uwerr for details).

Another useful check is also to devide the full ensemble into two (or more) sub-ensembles (using the parameter nrep). The summary function will then also compute a Q-value, which should be larger than \(0.1\) at least for the error to be trustworthy. Otherwise the autocorrlation times are presumably too large to determine the error reliably.

Finally it is also usefull to check the value of \(\chi\) per data point. Such a plot is produced when plot is used.

The data must be ordered as in the output of readcmicor, see help(readcmicor) for details.

The expected order of gamma matrices and operators (local-local, local-fuzzed, fuzzed-local and fuzzed-fuzzed) (fuzzed = non-local) is as follows for all charged mesons:

1) the 4 operators for each type must be sorted like local-local, local-fuzzed, fuzzed-local and fuzzed-fuzzed. (fuzzed=non-local)

2) The 20 available types must be in the following order:
order PP PA AP AA 44 P4 4P A4 4A for pion like \(P=\gamma_5\) \(A=\gamma_4\gamma_5\) \(4=\gamma_4\)
order 44 VV AA 4V V4 4A A4 VA AV for rho-a1 like \(4=\gamma_i\gamma_4\) \(V=\gamma_i\) \(A=\gamma_i\gamma_5\)
order BB SS - total 20 \(\gamma_i\gamma_4\gamma_5\)
\(S=I\)

In this routine only PP PA AP AA 44 P4 4P A4 4A are used. See also cmicor!

(cases with space index "i" are summed over i=1,2,3) (best choice is weaker coupling at sink - ie PA rather than AP
order of magnitudes \(P > 4 > A\) (4 mixes A)
order of magnitudes \(4\sim A > V\) (A mixes V))

itype=21 is conserved vector current at sink, \(\gamma_5\) at source ( iobs is LV 1, FV 5 )

pion will perform a fit of of the following matrix

PL	PF	AL	AF	4L	4F

PL	p1 p1 cosh	p1 p2 cosh	p1 p3 sinh	p1 p4 sinh	p1 p5 sinh	p1 p6 sinh

PF	p2 p1 cosh	p2 p2 cosh	p2 p3 sinh	p2 p4 sinh	p2 p5 sinh	p2 p6 sinh

AL	p3 p1 sinh	p3 p2 sinh	p3 p3 cosh	p3 p4 cosh	p3 p5 cosh	p3 p6 cosh

AF	p4 p1 sinh	p4 p2 sinh	p4 p3 cosh	p4 p4 cosh	p4 p5 cosh	p4 p6 cosh

4L	p5 p1 sinh	p5 p2 sinh	p5 p3 cosh	p5 p4 cosh	p5 p5 cosh	p5 p6 cosh

4F	p6 p1 sinh	p6 p2 sinh	p6 p3 cosh	p6 p4 cosh	p6 p5 cosh	p6 p6 cosh

PL	PF	AL	AF	4L

for coupling parameter \(p_1\) to \(p_6\) and a mass \(m_\mathrm{PS}\) entering \(\cosh\) (and \(\sinh\) in the same way) as \(\cosh(-m_\mathrm{PS}(T/2-t))\). The values of \(t\) are running from t1 to t2 as specified by the user.

Corresponding to the value of matrix.size only a submatrix of the matrix given above is fitted. Moreover, if no.masses larger than one additional masses and coupling parameters are introduced.

The pion decay constant is computed from the relation $$f_\mathrm{PS} = 4\kappa\mu\frac{p_1}{\sqrt{m_\mathrm{PS}}^3}$$ and the PCAC mass (for matrix.size > 2) from $$m_\mathrm{PCAC} = m_\mathrm{PS} \frac{p_3}{2p_1}$$ when the summary.pionfit function is called. For matrix.size > 4 also \(Z_V\) is computed as $$Z_\mathrm{V}=\frac{2\mu}{m_\mathrm{PS}}\frac{p_1}{p_5}$$

See also

Examples

library(hadron)
# NOT RUN {
cmicor <- readcmicor("pion.dat")
# }# NOT RUN {
pionfit <- pion(cmicor, kappa=0.160856, mu=0.004, t1=10, t2=23, no.masses=1, matrix.size=4)
# }# NOT RUN {
summary(pionfit)
# }