Plot Command For Class Ouputdata

Generic plot routine for class “ouputdata”. Currently it plots the plaquette history and the history of \(\Delta H\)

plot(data, skip=0, ...)

Arguments

data
object of class “outputdata” obtained from a read with readoutputdata

skip
number of trajectories to be skipped in analysis for plaquette and \(\exp(-\Delta H)\).

...
additional arguments passed to the generic plot function.

Value

list containing the “data”, an object of class “uwerr” called “plaq.res” containing the statisical analysis for the plaquette and a second object of type “uwerr” called “dH.res” with the statisical analysis for \(\exp(-\Delta H)\).

See also

Examples

# NOT RUN {
plaq <- readoutputdata("output.data")
# }# NOT RUN {
plaq.plot <- plot(plaq, skip=100)
# }# NOT RUN {
summary(plaq.plot$plaq.res)
# }