read disconnected loops from binary files

Reads disconnected loops from binary files.

readbinarydisc(files, T=48, obs=5, endian="little",
               excludelist=c(""), nrSamples, path="")

Arguments

files
list of filenames to be read. Can be created for instance using getorderedfilelist. The filelist is assumed to be ordered with number of samples running fastest, and the next to fastest nubmer of gauges.

T
time extend of correlation functions.

obs
each file may contain T*obs correlation functions. With obs one choses which observable to read in.

endian
the endianess of the binary file.

excludelist
files to exclude from reading.

nrSamples
the number of samples

path
path to be prepended to every filename.

Value

returns a list with two arrays cf and icf with real and imaginary parts of the loops, and integers Time, nrStypes=1, nrSamples and nrObs=1. Both of the arrays have dimension c(T, N), where N is the number of measurements (gauges) and T the time extend. Time is the time extend, nrStypes the number of smearing levels and nrObs the number of operators, both of which are currently fixed to 1.

Details

It is assumend that each file contains O*T complex doubles, where T is the time extend and O the number of observables in the file. It is assumed that complex is the fastest running index, next time and then observables. The different samples are assumend to be in different files. The file list is assumed to be ordered with number of samples running fastest, and then number of gauges.

See also

Examples

# NOT RUN {
files <- character()
# }# NOT RUN {
for(i in seq(600,1744,8)) files <- c(files, "C2_dis_u_conf", sprintf("%.04d", i), ".dat", sep=""))
# }
# NOT RUN {
cf <- readbinarydisc(files, obs=4, excludelist=c("C2_pi0_conf0632.dat"))
# }