Read Single Data Files in Chris Michael Format

reads data from single files in Chris Michael format

readcmicor(filename,
           colClasses=c("integer","integer","integer","numeric","numeric","integer"), skip=0)

readcmifiles(files, excludelist=c(""), skip, verbose=FALSE,
             colClasses, obs=NULL, obs.index)

readcmidatafiles(files, excludelist=c(""), skip=1, verbose=FALSE,
                 colClasses=c("integer","integer","integer","numeric","numeric"),
                 obs=NULL, obs.index=1)

readcmiloopfiles(files, excludelist=c(""), skip=0, verbose=FALSE,
                 colClasses=c("integer", "integer","integer","integer",
                              "numeric","numeric","numeric","numeric"),
                 obs=NULL, obs.index=2)

Arguments

filename
A single file name to read from.

files
list of filenames to be read. Can be created using getorderedfilelist.

skip
Number of lines to be skipped at the beginning of each file

excludelist
files to exclude from reading.

verbose
Increases verbosity of the function.

colClasses
The column data type classes, the read.table.

obs
To reduce memory consumption it is possible to extract only one of the observales. The column in which to match obs is to be given with obs.index. This will only be affective if obs is not NULL.

obs.index
The column in which to match obs is to be given with obs.index.

Value

readcmicor returns an object of class cmicor, read from a single file.

readcmidatafiles returns an object of class cmicor, which is an rbind of all data.frames read from the single files in the filelist.

readcmiloopfiles returns an object of class cmiloop, which is an rbind of all data.frames read from the single files in the filelist.

Details

These functions reads data from single data files. It is assumed that every file has the same number of columns.

The cmi (Chris Michael) format for connected correlators comprises 6 colums per file: 1) the observable type number (itype); 2) the operator type number (iobs); 3) the time difference from source going from 0 to \(T/2\) for each operator type; 4) \(c_1\) correlator value at time value forward in time; 5) \(c_2\) correlator value at time value backward in time; 6) number of gauge configuration.

There are scripts shipped with the package converting the output written into seperate files for each gauge configuration into the expected format. They are called puttogether.sh and puttogether_reverse.sh which will sort with increasing and with decreasing gauge configuration number, respectively.

Note, that the normalisation of correlators needs multiplication by factor of \(0.5\) (and possible \((2*\kappa)^2\) and \(L^3\) factors dependent on your conventions).

The values of itype run from 1 to the total number of gamma matrix combinations available. iobs equals 1 for local-local correlators, 3 for local-smeared, 5 for smeared-local and 7 for smeared-smeared

For charged mesons the order of gamma-matrix combinations is as follows:
order PP PA AP AA 44 P4 4P A4 4A for pion like \(P=\gamma_5\) \(A=\gamma_4\gamma_5\) \(4=\gamma_4\)
order 44 VV AA 4V V4 4A A4 VA AV for rho-a1 like \(4=\gamma_i\gamma_4\) \(V=\gamma_i\) \(A=\gamma_i\gamma_5\)
order BB SS - total 20 \(\gamma_i\gamma_4\gamma_5\) \(S=I\)
itype=21 is conserved vector current at sink, \(\gamma_5\) at source

For neutral mesons the order of gamma-matrix combinations is as follows:
order PP PA AP AA II PI IP AI IA for pion like \(P=\gamma_5\) \(A=\gamma_4\gamma_5\) \(I=1\)
order 44 VV BB 4V V4 4B B4 VB BV for rho-b1 like \(4=\gamma_i\gamma_4\) \(V=\gamma_i\) \(B=\gamma_i\gamma_4\gamma_5\)
order XX AA - total 20 for a0-X like \(A=\gamma_i\gamma_5\) \(X=\gamma_4\)

For loops (disconnected contributions to neutral mesons) the convention is as follows: files are assumed to have eight columns with gauge, gamma, t, sample, ReTL, ImTL, ReTF, ImTF, where gamma is 1 to 16 as list of (hermitian) gamma matrices: order g_5 g_1 g_2 g_3
-ig_4* g_5 g_1 g_2 g_3
-ig_5* i*g_5 g_1 g_2 g_3 ie 1,..
-ig_5g_4* -i*g_5 g_1 g_2 g_3 ie g_4, g_5*row 2
(so P is 1; A4 is 5; S is 9; A_i is 10,11,12 etc)

t is t-value of trace (here spatial momentum is zero) sample is sample number 1,...24 (or 96) ReTL is real part of trace at time t, with gamma combination given and Local operator (F is Fuzzed == non-local) operator).

Normalisation is trace M^-1 with M=1+...

To make a disconnected correlator, one combines these traces for different t (and different sample number) as a product. Note only Re Gamma=1 and Im Gamma=gamma_5 have VEV's, see computeDisc

See also

Examples

library(hadron)
# NOT RUN {
filelist <- getorderedfilelist("ouptrc")
# }# NOT RUN {
cmicor <- readcmidatafiles(filelist, skip=1)
# }