Read Gradient Flow Output Files in tmLQCD format

given a pathname, reads all gradient flow output files in that directory

readgradflow(path, skip=0, basename="gradflow", col.names)

Arguments

path
the path into which the function should descend

skip
number of measurements to skip.

basename
basename of the files to be read.

col.names
column names of the columns in the files to be read. If not given it will be infered from the files, if possible.

Value

The function returns a data frame ordered first by the flow time and then by the the trajectory number (so the trajectory number is the index which runs fastest). The data frame has column names

t - flow time

traj - trajectory number

P - plaquette expectation value (at flow time t)

Eplaq - energy density from plaquette definition (at flow time t)

Esym - energy density from clover definition (at flow time t)

tsqEplaq - flow time squared multiplied by plaquette energy density

tsqEsym - flow time squared multiplied by clover energy density

Wsym - BMW 'w(t)' observable

.

Details

This function reads all tmLQCD gradient flow files in the given path and returns a data frame which concatenates them all.

The single files are expected to be in the tmLQCD format which consists of a header with the column names "traj t P Eplaq Esym tsqEplaq tsqEsym Wsym" and the measurement for each flow time in rows. The columns can be ordered arbitrarily as long as the header and the data are consistent.

Examples

library(hadron)
# NOT RUN {
raw.gf <- readgradflow(path)
# }