Read correlator data from single file

Reads arbitrary number of samples for a complex correlation function from a text file.

readtextcf(file, T=48, sym=TRUE, path="", skip=1, check.t=0, ind.vector=c(2,3), symmetrise=TRUE,
                       sparsity=1, avg=1, Nmin=4, autotruncate=TRUE)

Arguments

file
filename of file to read from.

T
time extend of the correlation function

sym
if TRUE average C(+t) and C(-t), otherwise C(+t) and -C(-t). Averaging can be switched off using the symmetrise option.

skip
number of lines to skip at beginning of file

check.t
if set to an integer value larger than zero the function will assume that in the corresponding column of the file the Euclidean time is counted and it will check whether the maximum in this column is identical to T-1.

ind.vector
index vector of length 2 with the indices of real and imaginary values of correlator, respectivley.

symmetrise
if set to TRUE, the correlation function will be averaged for t and T-t, with the sign depending on the value of sym. Note that currently the correlator with t-values larger than T/2 will be discarded.

path
the path to the files.

autotruncate

Value

returns a list with two arrays cf and icf with real and imaginary parts of the correlator, and integers Time, nrStypes=1 and nrObs=1. Both of the arrays have dimension c(N, (Time/2+1)), where N is the number of measurements (gauges). Time is the time extend, nrStypes the number of smearing levels and nrObs the number of operators, both of which are currently fixed to 1.

See also