paste a number with error in tex-ready format

A number with error is converted to a string in tex-ready format like xx(yy) thereby automatically determining the digit at which the error applies.

tex.catwitherror(x, dx, digits=1, with.dollar=TRUE, human.readable=TRUE)

Arguments

x
either a single numeric value, or a numeric vector, where the first element is the value and the second is its error

dx
the error. If supplied, it will be printed as the error and the value is the first element of x. If dx is missing, the second element of x, if available, is used as the error. If dx is missing and the length of x is one, only the value is converted to a string without error.

digits
number of error digits

with.dollar
include the tex dollar in the return string or not

human.readable
controls whether or not scientific format will be produced in a human readable form, or not at all. The latter might be useful for output with summary or print.

Value

writes a string to standard output

Details

The value of the first element of x is properly rounded to its significant digits determined by the values of dx or the second element of x (see above) and digits. Then a tex-ready string is returned.

Examples

tex.catwitherror(x=0.375567, dx=0.001)

#> [1] "$0.376(1)$"

tex.catwitherror(x=c(0.375567, 0.001))

#> [1] "$0.376(1)$"

## it can be used with apply
x = array(c(0.1187, 0.291, 0.388, 0.011, 0.037, 0.021), dim=c(3,2))
apply(x, 1, tex.catwitherror, digits=2)

#> [1] "$0.119(11)$" "$0.291(37)$" "$0.388(21)$"