variational analysis of a correlator matrix

variational analysis of a correlator matrix

variational(Cor, ta, tb, tmax, N, T1, matrix.size, no.masses)

Arguments

Cor
correlator matrix

ta
first time value for eigenvector determination

tb
second time value for eigenvector determination

tmax
maximal time value to be considered in the analysis

N
matrix size. must be <= to the size of the correlator matrix

T1
maximal time value in correlator matrix

matrix.size
matrix size to be used for start value determination (e.g. for pionfit)

no.masses
number of mass values for start value determination

Value

returns a list with following entries

t

the list of time values

res.values

the mass values in a tmax-ta times N array for all analysed t-values

par

startvalues for a fit

variational.masses

the list of mass values as determined from the variational analysis

See also