Weighted Median Value

Compute a weighted median of a numeric vector.

weighted.median(x, w, na.rm=TRUE, ties=NULL)

Arguments

x
a numeric vector containing the values whose weighted median is to be computed.

w
a vector of weights the same length as x giving the weights to use for each element of x. Default value is equal weight to all values.

na.rm
a logical value indicating whether NA values in x should be stripped before the computation proceeds.

ties
a character string specifying how to solve ties between two x's that are satisfying the weighted median criteria. Note that at most two values can satisfy the criteria. When ties is "min", the smaller value of the two is returned and when it is "max", the larger value is returned. If ties is "mean", the mean of the two values is returned and if it is "both", both values are returned. Finally, if ties is "weighted" (or NULL) a weighted average of the two are returned, where the weights are weights of all values x[i] <= x[k] and x[i] >= x[k], respectively. Default value is NULL.

Value

Returns the weighted median.

Details

For the n elements x = c(x[1], x[2], ..., x[n]) with positive weights w = c(w[1], w[2], ..., w[n]) such that sum(w) = S, the weighted median is defined as the element x[k] for which initial the total weight of all elements x[i] < x[k] is less or equal to S/2 and for which the total weight of all elements x[i] > x[k] is less or equal to S/2 (c.f. [1]).

If w is missing then all elements of x are given the same positive weight. If all weights are zero, NA is returned.

When all the weights are the same and ties is "weighted" (or NULL) weighted.median gives the same result as median.

If one or more weights are Inf, it is the same as these weights have the same weight and the others has zero. This makes things easier for cases where the weights are result of a division with zero.

The weighted median solves the following optimization problem:

$$\alpha^* = \arg_\alpha \min \sum_{k=1}{K} w_k |x_k-\alpha|$$ where \(x=(x_1,x_2,\ldots,x_K)\) are scalars and \(w=(w_1,w_2,\ldots,w_K)\) are the corresponding "weights" for each individual \(x\) value.

See also

References

[1] T.H. Cormen, C.E. Leiserson, R.L. Rivest, Introduction to Algorithms, The MIT Press, Massachusetts Institute of Technology, 1989.

Examples

   x <- 1:10
   n <- length(x)
   median(x)                            # 5.5

#> [1] 5.5

   weighted.median(x)                   # 5.5

#> [1] 5.5

   w <- rep(1, n)
   weighted.median(x, w)                # 5.5 (default)

#> [1] 5.5

   weighted.median(x, ties="weighted")  # 5.5 (default)

#> [1] 5.5

   weighted.median(x, ties="min")       # 5

#> [1] 5

   weighted.median(x, ties="max")       # 6

#> [1] 6

   # Pull the median towards zero
   w[1] <- 5
   weighted.median(x, w)                # 3.5

#> [1] 3.5

   y <- c(rep(0,w[1]), x[-1])           # Only possible for integer weights
   median(y)                            # 3.5

#> [1] 3.5

   # Put even more weight on the zero
   w[1] <- 8.5
   weighted.median(x, w)                # 2

#> [1] 2

   # All weight on the first value
   w[1] <- Inf
   weighted.median(x, w)                # 1

#> [1] 1

   # All weight on the first value
   w[1] <- 1
   w[n] <- Inf
   weighted.median(x, w)                # 10

#> [1] 10

   # All weights set to zero
   w <- rep(0, n)
   weighted.median(x, w)                # NA

#> [1] NA